Изменил Новости
TODO: Сделать тесты новостей
